Fixed off-by-one pixel strip error
Mirroring the same merge request as the main godot-rust project. The clip path for group "g20" which holds the upper border of the claws was aligned exactly to the inside of the outer border. Due to rounding error, an orange strip was sometimes erroneously drawn where the two borders meet for each claw. This is especially noticeable in the godot-ferris icon at the top of the page. Expanding the bounds of the clip path is sufficient to alleviate the issue entirely.

update to show minor changes to image